add getAllRoles-Method, restructure ordner structure
