+++ backend/src/main/java/capstone/kpiboard/controller/KpiController.java
@@ -3,7 +3,7 @@ package capstone.kpiboard.controller;
 import capstone.kpiboard.exceptions.KpiNotFoundException;
 import capstone.kpiboard.model.kpi.Kpi;
 import capstone.kpiboard.model.kpi.NewKpi;
-import capstone.kpiboard.service.KpiService;
+import capstone.kpiboard.service.kpi.KpiService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
